* updated coc shaders

+++ res/gamedata_COC/shaders/r2/common.h
@@ -41,28 +41,28 @@
 #define parallax float2(PARALLAX_H, -PARALLAX_H/2)
 
 #ifdef        USE_R2_STATIC_SUN
-#  define xmaterial half(1.0h/4.h)
+#  define xmaterial float(1.0h/4.h)
 #else
-#  define xmaterial half(L_material.w)
+#  define xmaterial float(L_material.w)
 #endif
 //////////////////////////////////////////////////////////////////////////////////////////
-uniform half4                hemi_cube_pos_faces;
-uniform half4                hemi_cube_neg_faces;
-uniform half4                L_material;                            // 0,0,0,mid
-uniform half4                Ldynamic_color;                      // dynamic light color (rgb1)        - spot/point
-uniform half4                Ldynamic_pos;                       // dynamic light pos+1/range(w) - spot/point
-uniform half4                Ldynamic_dir;                        // dynamic light direction         - sun
+uniform float4                hemi_cube_pos_faces;
+uniform float4                hemi_cube_neg_faces;
+uniform float4                L_material;                            // 0,0,0,mid
+uniform float4                Ldynamic_color;                      // dynamic light color (rgb1)        - spot/point
+uniform float4                Ldynamic_pos;                       // dynamic light pos+1/range(w) - spot/point
+uniform float4                Ldynamic_dir;                        // dynamic light direction         - sun
 
-uniform half4                J_direct        [6];
-uniform half4                J_spot                [6];
+uniform float4                J_direct        [6];
+uniform float4                J_spot                [6];
 
-half          calc_fogging               (half4 w_pos)      { return dot(w_pos,fog_plane);         }
-half2         calc_detail                (half3 w_pos)      {
+float          calc_fogging               (float4 w_pos)      { return dot(w_pos,fog_plane);         }
+float2         calc_detail                (float3 w_pos)      {
         float                 dtl        = distance                (w_pos,eye_position)*dt_params.w;
                               dtl        = min              (dtl*dtl, 1);
-        half                  dt_mul     = 1  - dtl;        // dt*  [1 ..  0 ]
-        half                  dt_add     = .5 * dtl;        // dt+  [0 .. 0.5]
-        return                half2      (dt_mul,dt_add);
+        float                  dt_mul     = 1  - dtl;        // dt*  [1 ..  0 ]
+        float                  dt_add     = .5 * dtl;        // dt+  [0 .. 0.5]
+        return                float2      (dt_mul,dt_add);
 }
 float3         calc_reflection     (float3 pos_w, float3 norm_w)
 {
@@ -140,9 +140,9 @@ struct         p_bumped        {
         float2            tcdh        : TEXCOORD0;        // Texture coordinates
 #endif
         float4      position        : TEXCOORD1;        // position + hemi
-        half3       M1                : TEXCOORD2;        // nmap 2 eye - 1
-        half3       M2                : TEXCOORD3;        // nmap 2 eye - 2
-        half3       M3                : TEXCOORD4;        // nmap 2 eye - 3
+        float3       M1                : TEXCOORD2;        // nmap 2 eye - 1
+        float3       M2                : TEXCOORD3;        // nmap 2 eye - 2
+        float3       M3                : TEXCOORD4;        // nmap 2 eye - 3
 #ifdef USE_TDETAIL
         float2      tcdbump     	: TEXCOORD5;        // d-bump
     #ifdef USE_LM_HEMI
@@ -163,7 +163,7 @@ struct         p_flat                  {
     float2                    tcdh        : TEXCOORD0;        // Texture coordinates
 #endif
         float4                position        : TEXCOORD1;        // position + hemi
-        half3                N                : TEXCOORD2;        // Eye-space normal        (for lighting)
+        float3                N                : TEXCOORD2;        // Eye-space normal        (for lighting)
   #ifdef USE_TDETAIL
         float2                tcdbump                : TEXCOORD3;        // d-bump
     #ifdef USE_LM_HEMI
@@ -178,9 +178,9 @@ struct         p_flat                  {
 
 //////////////////////////////////////////////////////////////////////////////////////////
 struct                  f_deffer        		{
-        half4           position        		: COLOR0;        // px,py,pz, m-id
-        half4           Ne                		: COLOR1;        // nx,ny,nz, hemi
-        half4       	C                		: COLOR2;        // r, g, b,  gloss
+        float4           position        		: COLOR0;        // px,py,pz, m-id
+        float4           Ne                		: COLOR1;        // nx,ny,nz, hemi
+        float4       	C                		: COLOR2;        // r, g, b,  gloss
 };
 //////////////////////////////////////////////////////////////////////////////////////////
 struct  				p_screen                {
@@ -228,17 +228,17 @@ uniform sampler         s_image;                // used in various post-processi
 uniform sampler2D       s_tonemap;              // actually MidleGray / exp(Lw + eps)
 //////////////////////////////////////////////////////////////////////////////////////////
 // Defines                                		//
-#define def_gloss       half(2.f /255.f)
-#define def_aref        half(200.f/255.f)
-#define def_dbumph      half(0.333f)
-#define def_virtualh    half(0.05f)              // 5cm
-#define def_distort     half(0.05f)             // we get -0.5 .. 0.5 range, this is -512 .. 512 for 1024, so scale it
-#define def_hdr         half(9.h)         		// hight luminance range half(3.h)
-#define def_hdr_clip	half(0.75h)        		//
+#define def_gloss       float(2.f /255.f)
+#define def_aref        float(200.f/255.f)
+#define def_dbumph      float(0.333f)
+#define def_virtualh    float(0.05f)              // 5cm
+#define def_distort     float(0.05f)             // we get -0.5 .. 0.5 range, this is -512 .. 512 for 1024, so scale it
+#define def_hdr         float(9.h)         		// hight luminance range float(3.h)
+#define def_hdr_clip	float(0.75h)        		//
 
 //////////////////////////////////////////////////////////////////////////////////////////
-#define	LUMINANCE_VECTOR                 half3(0.3f, 0.38f, 0.22f)
-void        tonemap              (out half4 low, out half4 high, half3 rgb, half scale)
+#define	LUMINANCE_VECTOR                 float3(0.3f, 0.38f, 0.22f)
+void        tonemap              (out float4 low, out float4 high, float3 rgb, float scale)
 {
         rgb     =      	rgb*scale       ;
 
@@ -252,8 +252,8 @@ void        tonemap              (out half4 low, out half4 high, half3 rgb, half
 
         high	=		low/def_hdr		;        // 8x dynamic range
 #else
-        low		=       half4           ( ( (rgb*(1+rgb/fWhiteIntensitySQR)) / (rgb+1) ),           0 )	;
-        high	=       half4       	(rgb/def_hdr,   0 )	;		// 8x dynamic range
+        low		=       float4           ( ( (rgb*(1+rgb/fWhiteIntensitySQR)) / (rgb+1) ),           0 )	;
+        high	=       float4       	(rgb/def_hdr,   0 )	;		// 8x dynamic range
 #endif
 
 /*
@@ -263,42 +263,42 @@ void        tonemap              (out half4 low, out half4 high, half3 rgb, half
 	high	=	low/def_hdr;	// 8x dynamic range
 */
 
-//		low		= 	half4	(rgb, 0);
+//		low		= 	float4	(rgb, 0);
 //		rgb		/=	def_hdr	;
-//		high	= 	half4	(rgb, dot(rgb,0.333f)-def_hdr_clip)		;
+//		high	= 	float4	(rgb, dot(rgb,0.333f)-def_hdr_clip)		;
 }
-half4		combine_bloom        (half3  low, half4 high)	{
-        return        half4(low + high*high.a, 1.h);
+float4		combine_bloom        (float3  low, float4 high)	{
+        return        float4(low + high*high.a, 1.h);
 }
 
 float3	v_hemi        	(float3 n)                        	{        return L_hemi_color*(.5f + .5f*n.y);                   }
 float3	v_hemi_wrap     (float3 n, float w)                	{        return L_hemi_color*(w + (1-w)*n.y);                   }
 float3	v_sun           (float3 n)                        	{        return L_sun_color*dot(n,-L_sun_dir_w);                }
 float3	v_sun_wrap      (float3 n, float w)                	{        return L_sun_color*(w+(1-w)*dot(n,-L_sun_dir_w));      }
-half3   p_hemi          (float2 tc)                         {
-//        half3        	t_lmh         = tex2D             	(s_hemi, tc);
+float3   p_hemi          (float2 tc)                         {
+//        float3        	t_lmh         = tex2D             	(s_hemi, tc);
 //        return  dot     (t_lmh,1.h/4.h);
-        half4        	t_lmh         = tex2D             	(s_hemi, tc);
+        float4        	t_lmh         = tex2D             	(s_hemi, tc);
         return t_lmh.a;
 }
 
-half   get_hemi( half4 lmh)
+float   get_hemi( float4 lmh)
 {
 	return lmh.a;
 }
 
-half   get_sun( half4 lmh)
+float   get_sun( float4 lmh)
 {
 	return lmh.g;
 }
 
 //	contrast function
-half Contrast(half Input, half ContrastPower)
+float Contrast(float Input, float ContrastPower)
 {
      //piecewise contrast function
      bool IsAboveHalf = Input > 0.5 ;
-     half ToRaise = saturate(2*(IsAboveHalf ? 1-Input : Input));
-     half Output = 0.5*pow(ToRaise, ContrastPower);
+     float ToRaise = saturate(2*(IsAboveHalf ? 1-Input : Input));
+     float Output = 0.5*pow(ToRaise, ContrastPower);
      Output = IsAboveHalf ? 1-Output : Output;
      return Output;
 }
