+++ res/gamedata_COC/shaders/r2/shadow.h
@@ -10,7 +10,7 @@ uniform	sampler	s_smap	: register(ps,s0);	// 2D/cube shadowmap
 //////////////////////////////////////////////////////////////////////////////////////////
 // software
 //////////////////////////////////////////////////////////////////////////////////////////
-half 	sample_sw	(float2 tc, float2 shift, float depth_cmp)
+float 	sample_sw	(float2 tc, float2 shift, float depth_cmp)
 {
 	static const float 	ts = KERNEL / float(SMAP_size);
 	tc 		+= 		shift*ts;
@@ -26,15 +26,15 @@ half 	sample_sw	(float2 tc, float2 shift, float depth_cmp)
 		depth_cmp-tex2D	(s_smap, Tex01).x,
 		depth_cmp-tex2D	(s_smap, Tex10).x,
 		depth_cmp-tex2D	(s_smap, Tex11).x);
-	half4 	compare = step	(depth,0);
+	float4 	compare = step	(depth,0);
 	float2 	fr 		= frac	(Tex00*texsize);
-	half2 	ifr 	= half2	(1,1) - fr;
-	half4 	fr4 	= half4	(ifr.x*ifr.y, ifr.x*fr.y, fr.x*ifr.y,  fr.x*fr.y);
+	float2 	ifr 	= float2	(1,1) - fr;
+	float4 	fr4 	= float4	(ifr.x*ifr.y, ifr.x*fr.y, fr.x*ifr.y,  fr.x*fr.y);
 	return	dot		(compare, fr4);
 }
-half 	shadow_sw	(float4 tc)	{ 
+float 	shadow_sw	(float4 tc)	{ 
 	float2	tc_dw	= tc.xy / tc.w;
-	half4	s;
+	float4	s;
 	s.x	= sample_sw	(tc_dw,float2(-1,-1),tc.z); 
 	s.y	= sample_sw	(tc_dw,float2(+1,-1),tc.z); 
 	s.z	= sample_sw	(tc_dw,float2(-1,+1),tc.z); 
@@ -45,7 +45,7 @@ half 	shadow_sw	(float4 tc)	{
 //////////////////////////////////////////////////////////////////////////////////////////
 // hardware + PCF
 //////////////////////////////////////////////////////////////////////////////////////////
-half  	sample_hw_pcf	(float4 tc,float4 shift){
+float  	sample_hw_pcf	(float4 tc,float4 shift){
 	static const float 	ts = KERNEL / float(SMAP_size);
 #ifndef SUNSHAFTS_DYNAMIC
 	return tex2Dproj	(s_smap,tc + tc.w*shift*ts).x;
@@ -55,11 +55,11 @@ half  	sample_hw_pcf	(float4 tc,float4 shift){
 	return tex2Dlod(s_smap, tc2);
 #endif	//	SUNSHAFTS_DYNAMIC
 }
-half 	shadow_hw	(float4 tc)		{
-  half  s0		= sample_hw_pcf	(tc,float4(-1,-1,0,0)); 
-  half  s1		= sample_hw_pcf	(tc,float4(+1,-1,0,0)); 
-  half  s2		= sample_hw_pcf	(tc,float4(-1,+1,0,0)); 
-  half  s3		= sample_hw_pcf	(tc,float4(+1,+1,0,0));
+float 	shadow_hw	(float4 tc)		{
+  float  s0		= sample_hw_pcf	(tc,float4(-1,-1,0,0)); 
+  float  s1		= sample_hw_pcf	(tc,float4(+1,-1,0,0)); 
+  float  s2		= sample_hw_pcf	(tc,float4(-1,+1,0,0)); 
+  float  s3		= sample_hw_pcf	(tc,float4(+1,+1,0,0));
 
   return	(s0+s1+s2+s3)/(4.h);
 }
@@ -69,7 +69,7 @@ half 	shadow_hw	(float4 tc)		{
 //////////////////////////////////////////////////////////////////////////////////////////
 
 /*
-half  	sample_hw_f4	(float4 tc,float4 shift){
+float  	sample_hw_f4	(float4 tc,float4 shift){
 	static const float 	ts 	= KERNEL / 	float(SMAP_size);
 	float4	D4				= tex2Dproj	(s_smap,tc + tc.w*shift*ts);
 	float4 	dcmp			= tc.z/tc.w	;
@@ -78,7 +78,7 @@ half  	sample_hw_f4	(float4 tc,float4 shift){
 }
 */
 
-half  	sample_hw_f4	(float4 tc,float4 shift){
+float  	sample_hw_f4	(float4 tc,float4 shift){
 	static const float 	ts 	= 	KERNEL / 	float(SMAP_size);
 	float4 	T4				= 	tc/tc.w		;
 			T4.xy			+=	shift.xy*ts	;
@@ -88,20 +88,20 @@ half  	sample_hw_f4	(float4 tc,float4 shift){
 
 	float  	texsize 		= 	SMAP_size	;
 	float2 	fr 				= 	frac		(T4.xy * texsize);
-	half2 	ifr 			= 	half2		(1,1) - fr;
-	half4 	fr4 			= 	half4		(ifr.x*ifr.y, ifr.x*fr.y, fr.x*ifr.y,  fr.x*fr.y);
-	half4 	fr4s		 	= 	fr4.zywx	;
+	float2 	ifr 			= 	float2		(1,1) - fr;
+	float4 	fr4 			= 	float4		(ifr.x*ifr.y, ifr.x*fr.y, fr.x*ifr.y,  fr.x*fr.y);
+	float4 	fr4s		 	= 	fr4.zywx	;
 
 	return	dot	(compare, fr4s)	;
 	// return 	dot	(compare, 1.h/4.h)	;
 }
 
 
-half 	shadow_hw_f4	(float4 tc)		{
-  half  s0	= sample_hw_f4	(tc,float4(-1,-1,0,0)); 
-  half  s1	= sample_hw_f4	(tc,float4(+1,-1,0,0)); 
-  half  s2	= sample_hw_f4	(tc,float4(-1,+1,0,0)); 
-  half  s3	= sample_hw_f4	(tc,float4(+1,+1,0,0));
+float 	shadow_hw_f4	(float4 tc)		{
+  float  s0	= sample_hw_f4	(tc,float4(-1,-1,0,0)); 
+  float  s1	= sample_hw_f4	(tc,float4(+1,-1,0,0)); 
+  float  s2	= sample_hw_f4	(tc,float4(-1,+1,0,0)); 
+  float  s3	= sample_hw_f4	(tc,float4(+1,+1,0,0));
 	return	(s0+s1+s2+s3)/4.h;
 }
 
@@ -114,20 +114,20 @@ uniform sampler2D	jitter0;
 uniform sampler2D	jitter1;
 uniform sampler2D	jitter2;
 uniform sampler2D	jitter3;
-uniform half4 		jitterS;
-half4 	test 		(float4 tc, half2 offset)
+uniform float4 		jitterS;
+float4 	test 		(float4 tc, float2 offset)
 {
 	float4	tcx	= float4 (tc.xy + tc.w*offset, tc.zw);
 	return 	tex2Dproj (s_smap,tcx);
 }
 
-half 	shadowtest 	(float4 tc, float4 tcJ)				// jittered sampling
+float 	shadowtest 	(float4 tc, float4 tcJ)				// jittered sampling
 {
-	half4	r;
+	float4	r;
 
 	const 	float 	scale 	= (2.7f/float(SMAP_size));
-	half4	J0 	= tex2Dproj	(jitter0,tcJ)*scale;
-	half4	J1 	= tex2Dproj	(jitter1,tcJ)*scale;
+	float4	J0 	= tex2Dproj	(jitter0,tcJ)*scale;
+	float4	J1 	= tex2Dproj	(jitter1,tcJ)*scale;
 
 		r.x 	= test 	(tc,J0.xy).x;
 		r.y 	= test 	(tc,J0.wz).y;
@@ -137,51 +137,51 @@ half 	shadowtest 	(float4 tc, float4 tcJ)				// jittered sampling
 	return	dot(r,1.h/4.h);
 }
 
-half 	shadowtest_sun 	(float4 tc, float4 tcJ)			// jittered sampling
+float 	shadowtest_sun 	(float4 tc, float4 tcJ)			// jittered sampling
 {
-	half4	r;
+	float4	r;
 
 	//	const 	float 	scale 	= (2.0f/float(SMAP_size));
 	const 	float 	scale 	= (0.7f/float(SMAP_size));
 	
 
 	float2 	tc_J	= frac(tc.xy/tc.w*SMAP_size/4.0f )*.5f;
-	half4	J0 	= tex2D	(jitter0,tc_J)*scale;
-	//half4	J1 	= tex2D	(jitter1,tc_J)*scale;
+	float4	J0 	= tex2D	(jitter0,tc_J)*scale;
+	//float4	J1 	= tex2D	(jitter1,tc_J)*scale;
 
 	const float k = .5f/float(SMAP_size);
-	r.x 	= test 	(tc, J0.xy+half2(-k,-k)).x;
-	r.y 	= test 	(tc, J0.wz+half2( k,-k)).y;
-	r.z		= test	(tc,-J0.xy+half2(-k, k)).z;
-	r.w		= test	(tc,-J0.wz+half2( k, k)).x;
+	r.x 	= test 	(tc, J0.xy+float2(-k,-k)).x;
+	r.y 	= test 	(tc, J0.wz+float2( k,-k)).y;
+	r.z		= test	(tc,-J0.xy+float2(-k, k)).z;
+	r.w		= test	(tc,-J0.wz+float2( k, k)).x;
 
 	return	dot(r,1.h/4.h);
 }
 
-half 	shadow_high 	(float4 tc)			// jittered sampling
+float 	shadow_high 	(float4 tc)			// jittered sampling
 {
 
 	const	float 	scale 	= (0.5f/float(SMAP_size));
 
 	float2 	tc_J	= frac(tc.xy/tc.w*SMAP_size/4.0f )*.5f;
-	half4	J0 	= tex2D	(jitter0,tc_J)*scale;
+	float4	J0 	= tex2D	(jitter0,tc_J)*scale;
 
 	const float k = 1.f/float(SMAP_size);
-	half4	r;
-	r.x 	= test 	(tc,J0.xy+half2(-k,-k)).x;
-	r.y 	= test 	(tc,J0.wz+half2( k,-k)).y;
+	float4	r;
+	r.x 	= test 	(tc,J0.xy+float2(-k,-k)).x;
+	r.y 	= test 	(tc,J0.wz+float2( k,-k)).y;
 	
- 	r.z		= test	(tc,J0.xy+half2(-k, k)).z;
- 	r.w		= test	(tc,J0.wz+half2( k, k)).x;
+ 	r.z		= test	(tc,J0.xy+float2(-k, k)).z;
+ 	r.w		= test	(tc,J0.wz+float2( k, k)).x;
 	
 
 	const float k1 = 1.3f/float(SMAP_size);
-	half4	r1;
-	r1.x 	= test 	(tc,-J0.xy+half2(-k1,0)).x;
-	r1.y 	= test 	(tc,-J0.wz+half2( 0,-k1)).y;
+	float4	r1;
+	r1.x 	= test 	(tc,-J0.xy+float2(-k1,0)).x;
+	r1.y 	= test 	(tc,-J0.wz+float2( 0,-k1)).y;
 
-	r1.z	= test	(tc,-2*J0.xy+half2( k1, 0)).z;
- 	r1.w	= test	(tc,-2*J0.wz+half2( 0, k1)).x;
+	r1.z	= test	(tc,-2*J0.xy+float2( k1, 0)).z;
+ 	r1.w	= test	(tc,-2*J0.wz+float2( 0, k1)).x;
 
 	return ( r.x + r.y + r.z + r.w + r1.x + r1.y + r1.z + r1.w )*1.h/8.h;
 }
@@ -191,41 +191,41 @@ half 	shadow_high 	(float4 tc)			// jittered sampling
 //////////////////////////////////////////////////////////////////////////////////////////
 #ifdef	USE_HWSMAP_PCF
 // D24X8+PCF
-	half 	shadow		(float4	tc)		{ return shadow_hw	(tc);	}
+	float 	shadow		(float4	tc)		{ return shadow_hw	(tc);	}
 #else
 	#ifdef USE_FETCH4
 		// DF24+Fetch4
-		half 	shadow 		(float4 tc)		{ return shadow_hw_f4(tc);	}
+		float 	shadow 		(float4 tc)		{ return shadow_hw_f4(tc);	}
 	#else
 		// FP32
-		half 	shadow		(float4 tc) 	{ return shadow_sw	(tc);	}
+		float 	shadow		(float4 tc) 	{ return shadow_sw	(tc);	}
 	#endif
 #endif
 
 
 #ifdef	USE_HWSMAP_PCF
 	// D24X8+PCF
-	half 	shadow_volumetric		(float4	tc)		{ return sample_hw_pcf	( tc, float4(0,0,0,0) ); }
+	float 	shadow_volumetric		(float4	tc)		{ return sample_hw_pcf	( tc, float4(0,0,0,0) ); }
 #else
 	#ifdef USE_FETCH4
 		// DF24+Fetch4
-		half 	shadow_volumetric 		(float4 tc)		{ return sample_hw_f4	(tc, float4(0,0,0,0)); }
+		float 	shadow_volumetric 		(float4 tc)		{ return sample_hw_f4	(tc, float4(0,0,0,0)); }
 	#else
 		// FP32
-		half 	shadow_volumetric 		(float4 tc) 	{ return sample_sw	(tc.xy / tc.w,float2(0,0),tc.z); }
+		float 	shadow_volumetric 		(float4 tc) 	{ return sample_sw	(tc.xy / tc.w,float2(0,0),tc.z); }
 	#endif
 #endif
 
 //////////////////////////////////////////////////////////////////////////////////////////
 #ifdef  USE_SUNMASK	
 uniform float3x4	m_sunmask	;				// ortho-projection
-half 	sunmask		(float4 P)	{				// 
+float 	sunmask		(float4 P)	{				// 
 	float2 		tc	= mul	(m_sunmask, P);		//
 	return 		tex2D 		(s_lmap,tc).w;		// A8 
 	
 }
 #else
-half 	sunmask		(float4 P)	{ return 1.h; }	// 
+float 	sunmask		(float4 P)	{ return 1.h; }	// 
 #endif
 
 //////////////////////////////////////////////////////////////////////////////////////////
